+++ klarinet-coroutines/src/androidMain/kotlin/com/vectencia/klarinet/coroutines/IoDispatcher.android.kt
@@ -1,0 +1,7 @@
+package com.vectencia.klarinet.coroutines
+
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.IO
+
+internal actual val defaultIoDispatcher: CoroutineDispatcher = Dispatchers.IO